Minor refactor of retriever tests to make them easier to understand and work with
